테스트케이스 추가 및 버그 수정3

+++ src/main/domain/ledger/event/service/event_service.py
@@ -1,3 +1,4 @@
+from fastapi import HTTPException, status
 from sqlalchemy.orm import Session
 
 from domain.ledger.event.dto import CreateEventDTO
@@ -6,6 +7,7 @@ from domain.ledger.event.dto.edit_event_dto import EditEventDto
 from domain.ledger.event.dto.event_response_dto import EventResponseDTO
 from domain.ledger.event.dto.search_event_params import SearchEventParams
 from domain.ledger.event.repository import EventRepository
+from domain.organization.organization.repository import OrganizationRepository
 
 
 class EventService:
@@ -13,14 +15,38 @@ class EventService:
     def __init__(
             self,
             event_repository: EventRepository,
+            organization_repository: OrganizationRepository,
 
     ):
         self.event_repository = event_repository
+        self.organization_repository = organization_repository
 
     async def create_event(self, db:Session, create_event_dto:CreateEventDTO):
+        # verify
+        if create_event_dto.end_date < create_event_dto.start_date:
+            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="End date cannot be before start date")
+        if create_event_dto.start_date.year != create_event_dto.year:
+            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,detail="Start year must be equal to year")
+        if create_event_dto.end_date.year != create_event_dto.year:
+            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="End year must be equal to year")
+        organization = await self.organization_repository.find_by_id(db, create_event_dto.organization_id)
+        if organization is None:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Organization not found")
+        if not organization.start_year <= create_event_dto.year <= organization.end_year:
+            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="not available year")
+
+        # work
         await self.event_repository.create_event(db, create_event_dto)
 
     async def find_all(self, db:Session, search_event_params:SearchEventParams):
+        # verify
+        organization = await self.organization_repository.find_by_id(db, search_event_params.organization_id)
+        if organization is None:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Organization not found")
+        if not organization.start_year <= search_event_params.year <= organization.end_year:
+            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="not available year")
+
+        # work
         events = await self.event_repository.find_all(
             db=db,
             organization_id=search_event_params.organization_id,
@@ -28,9 +54,32 @@ class EventService:
         return [EventResponseDTO.model_validate(event) for event in events]
 
     async def update(self, db:Session, edit_event_dto:EditEventDto):
-        event = await self.event_repository.find_event_by_id(db, id=edit_event_dto.event_id)
+        # verify
+        if edit_event_dto.end_date < edit_event_dto.start_date:
+            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="End date cannot be before start date")
+        event = await self.event_repository.find_by_organization_and_id(
+            db,
+            organization_id=edit_event_dto.organization_id,
+            event_id=edit_event_dto.event_id)
+        if event is None:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Event not found")
+
+        if edit_event_dto.start_date.year != event.year:
+            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,detail="Start year must be equal to year")
+        if edit_event_dto.end_date.year != event.year:
+            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="End year must be equal to year")
+
+        # work
         await self.event_repository.update(db, event, edit_event_dto)
 
     async def delete(self, db:Session, delete_event_dto:DeleteEventParams):
-        event = await self.event_repository.find_event_by_id(db, id=delete_event_dto.event_id)
+        # verify
+        event = await self.event_repository.find_by_organization_and_id(
+            db,
+            organization_id=delete_event_dto.organization_id,
+            event_id=delete_event_dto.event_id)
+        if event is None:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Event not found")
+
+        # work
         await self.event_repository.delete(db, event)
